chore(prettier): implement base configuration and update exports
ref #35
