+++ src/prettier/configs/index.ts
@@ -1,0 +1,3 @@
+import { base } from './base';
+
+export const configs = { base };
