+++ src/prettier/index.spec.ts
@@ -1,7 +1,7 @@
 import { prettier } from './index';
 
 describe('prettier', () => {
-  const expectedKeys = ['getConfig'] as const;
+  const expectedKeys = ['configs'] as const;
 
   it(`should have exactly ${expectedKeys.length} keys`, () => {
     expect(Object.keys(prettier).length).toBe(expectedKeys.length);
@@ -12,7 +12,7 @@ describe('prettier', () => {
   });
 
   it.each(expectedKeys)(
-    'should have "%s" as a valid function in prettier',
+    'should have "%s" as a valid object in prettier',
     (property) => {
       expect(prettier).toBeInstanceOf(Object);
       expect(prettier).toHaveProperty(property);
@@ -21,32 +21,4 @@ describe('prettier', () => {
       );
     },
   );
-
-  describe('getConfig', () => {
-    const expectedConfig = {
-      printWidth: 80,
-      semi: true,
-      tabWidth: 2,
-      trailingComma: 'all',
-      singleQuote: true,
-      bracketSpacing: true,
-      bracketSameLine: true,
-      endOfLine: 'lf',
-    };
-
-    const expectedObject: Record<string, unknown> = Object.fromEntries(
-      Object.keys(expectedConfig).map((key) => [key, expect.anything()]),
-    );
-
-    it('should return the exact configuration object', () => {
-      const config = prettier.getConfig();
-      expect(config).toStrictEqual(expectedConfig);
-    });
-
-    it('should return a valid object with all expected keys', () => {
-      const config = prettier.getConfig();
-      expect(config).toBeInstanceOf(Object);
-      expect(config).toMatchObject(expectedObject);
-    });
-  });
 });
